parsing fixes
